Attempt to include roll number for building society transactions
We will attempt to detect payments involving building society accounts
by checking the sort code and/or account number. If one is found, the
corresponding roll number pattern will be checked against the
appropriate field on the record to see if it does indeed contain
a roll number. If found it will be included in the uploaded
transaction.