Parse a wider range of WorldPay settlement reference date formats
WorldPay settlements can be in the form:
- "PREFIX -0101" with the last 4 digits being a day/month
- "PREFIX -01" with the last 2 digits being a day
- "PREFIX -" with no date reference

The 4-digit format was already handled and now a 2-digit date is also handled. Sometimes no date appears in the reference – these transactions should still be labelled as "administrative" rather than "unidentified".

TODO: should there be a cap on how far before the record's date the settlement batch date can be? If a batch seems to be, say, 1 week before the record, perhaps the two should not be connected?